Add clear button to files search
I would have preferred it to just depend on the default styling of
search type inputs in browsers, which at least on Webkit and IE will add
a clear button automatically, but Firefox doesn't add one and the
corresponding ticket [1] also doesn't look like that will change soon.

So a custom implementation it is. Hooray.

Implements #2983

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=558594