💥 Unbundling the cura plugin
It lives on as curalegacy at

  https://github.com/OctoPrint/OctoPrint-CuraLegacy

which will be added to the plugin repository once the first RC of 1.3.11
lands.

Due to this change, the general handling of STL files has also changed
slightly. They will no longer be supported by upload out of the box
unless a SlicerPlugin is available (which will now by default extend the
extension tree with STL files, unless overwritten). As a consequence of
that the file type filters in the file list have also been adjusted to
only show up if there are additional file types to filter for.