☔ Don't confuse users with pip's "OMG PY2.7 IS EOL!!1!" message
We know that python 2.7 is about to reach end of life. We are making
preparations for this in the devel branch. We don't need pip going
around and alarming our users about something they cannot solve, and
which pretty much is already solved anyhow.

So we nuke this stupid message present in current pip versions. Hard.

See also #2988