Allow subscriptions of multiple events
events > subscriptions[] > event may be either a string (existing
functionality) or now a list of strings.  If a list, the subscriptions
are treated the same for each event in the list.

To make it possible for command subscriptions to determine which event
was triggered, a {__eventname} substitution has been added.