🐛 self._user might not be a SessionUser
If the socket is already logged out it will be an AnonymousUser
and thus not a SessionUser.

We shouldn't actually be pushing this without a round over an RC but
frankly I don't want to postpone the release for two weeks (as next week
is kinda busy) just to solve this obvious issue with zero side effects.

So here I am doing things I shouldn't.