✨ Remove feed/flow rate upper limits altogether
Gets rid of sliders as well (since those don't work without an upper
limit) and adds a note to both that they are "write-only" as we don't
have a way to query the current values from the firmware (yay...).

Modifies #3503

Also fixes a bug in #3503, which only adjusted the client side
boundaries but forgot about the server side.