✨ Support parsing of M33 and M20 w/ long names
As to my knowledge no firmware currently implements M20 with long file
names like this:

  Begin file list
  UGLYNA~1.GCO 2342 Uglyname with some_longer-name.gcode
  End file list

However, as proposed by myself back in 2016 that would make the most
sense (contrary to M33, which requires n+1 requests to get all metadata
from the printer), so OctoPrint now supports it and firmwares are
welcome to implement it accordingly to finally implement a *usable* way
to retrieve long filenames that doesn't saturate the serial connection
for an undeterminable amount of time just to retrieve a list of file
names.

Compare MarlinFirmware/Marlin#4733.

Also add parsing of the (not really parser friendly) output of M33.

Note that OctoPrint will not send M33 on its own even though it now
supports parsing its response. There's no capability flag for this as
far as documented, and there appears to exist a GCODE clash according
to https://reprap.org/wiki/G-code, so just blindly sending M33 to the
printer from core might cause severe issues.

Users may instead add active querying for the long name (if not included
in the M20 response) through a plugin, as demonstrated in

  https://gist.github.com/foosel/cadc38123687a46816a1d3e1d160d48a

Closes #1600