✨ New recovery endpoint & refactoring
At /recovery/ there's now a (for now basic) recovery dialog that
currently allows the core functionality from the system command menu, so
shutting down and rebooting the system, restarting OctoPrint and
restarting OctoPrint in safe mode. The most important part here is the
ability to restart in safe mode, as that will allow to get OctoPrint
back into working order if an errant third party plugin or language
pack has made the standard UI non functional.

For that reason, no third party plugins are enabled or utilized on the
recovery dialog, even the theming hook is aggressively error resilient.

In order to still allow to authenticate the user (no one but ADMIN
should be allowed to use the recovery UI) the loginui plugin has been
merged into core as its own endpoint /login/ which coreui and recovery
can redirect to. This is a welcome change considering that the loginui
was absolutely necessary anyhow, and it being a plugin was merely a side
effect of backporting from forcelogin. This has now been cleaned up.