✨ Login dialog now supports permission check
That way we can display a login dialog if we require a specific set of
permissions, which then also gets displayed if the user is logged in but
lacks them.

If the user is already logged in and has the requested permissions, an
immediate redirect to the redirect URI is triggered.