🐛 Fix caching wrapper on endpoints
At some point the last modified implementation broke, and for some
reason current Firefox also appears to be even sending an
If-Modified-Since header if an ETag is present.

This fixes caching again so that if both If-Modified-Since and
If-None-Match are present, Last-Modified and ETag will be checked,
if onlye one is present only that one will be checked, and if neither
is present the cache will always be considered stale.

Possible (at least partial) solution for #1241