:arrow_up: Upgrade FontAwesome to 5.15.1
This is not a breaking change, due to the inclusion of the v4-shims. All icons should aim to use fontawesome 5 styles, full guide on breaking changes can be found at:

https://fontawesome.com/how-to-use/on-the-web/setup/upgrading-from-version-4

I've aimed to update all icons I could find, but due to the compatibility layer if there are a couple missed, they still all work fine.