✨ New system info collection, ready for copy/paste
Should hopefully make it easier in the future to collect data needed
for ticket processing. Can be accessed through System Information in the
About dialog (also directly linked from the footer and the settings),
via a new sub command `octoprint systeminfo` on the command
line and on the recovery page.

Currently included:

  * octoprint version
  * safe mode yes/no
  * printer firmware name (if connected)
  * browser user agent (if dump happens through API)
  * everything reported by the environment detector (that includes data
    from plugins as well that add to that, e.g. pi_support)
  * connectivity check information

Might be extended more in the future.