🥅 Prevent assignment of dangerous permissions to guests
If we don't, chances are high someone will happily assign the admin
group as a subgroup to the guest group and then publicly expose their
instance on the internet, thus completely defeating the purpose of
making access control mandatory.

We'll also automatically sanitize such configurations now but log a
warning. I don't see a better way to protect the innocent, sorry.