✨ New --overlay CLI flag
Allows to specify additional config overlay paths to use between the
default settings and config.yaml. Can be either individual yaml files,
or a whole directory containing an arbitrary number of yaml files.

Allows to change the default configuration without interfering with
config.yaml.

OctoPrint side solution for #3821, OctoPi needs to be modified to make
use of this for its platform specific configuration.