🐛 Better fix for folder listing on files API
We don't want children to be included on api/files or api/files/origin,
but we do want them on api/files/origin/path as that gets taken from
a file listing itself. So a level parameter is introduced on the storage
interface, defaulting to 0 and thus the former behaviour (remove
children), but setting it to a value > 0 will only remove grand children
and thus give us a way to query what we want on the detail endpoint.

See #3844