🐛 CLI: Force set values on the config CLI
We don't initialize the settings plugins (and we don't want to change
that either since it's a TON of potentially error prone overhead just
to allow a glorified editing of a yaml file), so we now can force
values even if their paths are unknown to OctoPrint. Can be abused to
clutter config.yaml, but so can plugins.

Closes #3869