✨ Tracking: add printer state to ping event
That will allow us to get a general idea of actively used vs idle
instances. A huge number of instances in an error state could also be an
indicator of problems with a release.

Would also be a cool thing to be able to add "this many printers are
currently printing" to data.octoprint.org