🐛 pmgr: make upload install async & fix conflict handling
It turns out that a) we were not yet doing plugin installation after an
archive/file upload asynchronously, so that any delays there could still
cause timeouts, and that b) we were also not handling the conflict
situation of trying to start another plugin install while one is already
running correctly on the frontend.

This fixes both issues.