🐛 Fix fancy/unfancy terminal switching
The switch back was enqueued in a timeout, however that timeout was only
ever cleared if fancy functionality was already enabled when trying to
switch it back on due to a logic error. Now it gets cleared whenever the
time needed for processing the log lines is above the threshold.

Potentially fixes #3905