🐛 backup: fix plugin repo access on cli
We were trying to fetch the repo url from the settings. However, due to
the settings overlays from other plugins not being available when
operating in a CLI environment, the repository URL was unset and thus
nothing was fetched.

This is now fixed by falling back to the default repo URL from the
pluginmanager, which has been extracted into a constant for that reason.

Fixes #3883