Use G0 for jogging
By convention `G1` is used for cutting/burning/extruding moves at normal speed, and `G0` is intended for rapid moves that do not invoke tooling. 
e.g. https://marlinfw.org/docs/gcode/G000-G001.html

Using G1 can have adverse effects in many environments. For example in Marlin the laser will be turned on for G1 moves and off for G1 (`LASER_MOVE_G0_OFF`) and it is easy to unintentionally turn on the laser when not wearing eye protection such as when moving to load material.