:pencil2: Show meaningful message when software update is blocked
And log when it is undervoltage.

The problem here was twofold: first, a very generic message asking to check the log confused users. Second, there was nothing logged when it said check the logs.

So, this changes that so the message is escaped and displayed in the pnotify and when the update is blocked for undervoltage this is logged.

This does not add logging for other methods, since it seems unnecessary to me.

Closes #4024