🔒 Make sure to validate log paths for zip bundle
Also do the same for timelapses.

For individual files, the LargeResponseHandler takes care of this
validation. For zip downloads however, we need to do it. An earlier
refactoring accidentally applied the same logic to the invididual
downloads as well, though in a broken way, this now fixes behaviour in
both variants.