✨ Confirm overwrite dialog
On file upload, if the target file already exists, a confirmation will
now be shown that allows to explicitly overwrite the existing file,
rename the file prior to upload (with a collision free suggestion
provided) or cancel the upload.

A feature flag allows to disable this confirmation again to restore the
old behaviour in case the new one clashes with people's workflows.

Implements #3873