:bug: Fix height of the webcam test modal
An interesting condition of when the correct positioning was calculated - with the animation, often the webcam stream loaded fast enough that it was before the animation was finished. Now there is no animation (from jQuery) then we hit a timing problem.

Yes, it is hardcoded but so is the modal width.