🥅 GCV: Work around for Chrome bug
Chrome's canvas arc method seems to produce wrong results in case of
extremely large radii combined with extremely small enclosed angles. So
in case of Chrome we'll now default to a replacement function
approximating the arc with bezier curves. This can however be disabled
via the advanced options.

Additionally, another debug visualization option was added that shows
the arcs with their center points and segment borders.

Closes #4117