✨ Send M20 L if detected as supported
Also support quoted long names, which apparently are sent by some
firmwares? Make a suggestion for a format that is universally parseable,
document it on the reprep wiki, people STILL find a way to be
incompatible to it. I can't even...

Anyhow. See https://community.octoprint.org/t/adding-support-for-m20-l/34402