💥 Deprecate OctoPi 0.15 environment
On OctoPi 0.15 OctoPrint 1.7.0+ will refuse to update.

Reason: yet another dependency no longer building in
this ancient environment. We'll allow an upgrade to 1.7.0
because we cannot retroactively warn people about this
anymore, but from here on out that's it.

OctoPi 0.15 was released in 2018, I think it's fair to
expect people to update, especially with the available
backup and restore functionality.