♻ Improved implementation of #4281
Instead of forcing plugins to use absolute paths to templates,
via a plugin aware jinja2 environment it is now detected whether
the template they are importing does locally exist in their own
template folder (in which case it gets automatically prefixed) or not.

A deprecation notice is now only logged when a plugin imports a
template from another template without the prefix (which probably
is a *very* rare occurence anyhow).