🐛 Ignore checks w/o type
If we disabled some checks in the past, and their
providing plugins are currently disabled, we
will have checks configured that only consist
of a "disabled" key. That will cause a warning
to be logged as there's no type defined. So
we now simply ignore entries without a type
altogether.