💥 Upgrade Flask & thus Jinja to fix warnings
Also iron out a bunch of other warnings.

Heads-up: This upgrades Jinja, werkzeug and itsdangerous as well!

Jinja 3 is not fully backwards compatible to 2
due to a bug fix we so far pinned to an old
version for. It's time to stop doing that
with OctoPrint 2.0.0. Should only affect plugins that
are still trying to access variables defined in
an outer scope in an inner scope, as deemed
an unintentional behaviour in pallets/jinja#641
and removed in Jinja 2.9. It's easy to solve
by switching to using assignments, see
https://jinja.palletsprojects.com/en/3.0.x/templates/#assignments.