Use json for annoucements cache
Parsing this XML file is very slow, so there's no reason to do it more
than we have to.

See 4480-2021-12-11T13:23:48-05:00.json, MainThread -> parse

On my system, this reduces the time for /plugin/announcements/channels
from 250ms CPU time to 50ms.

The rough improvment in this specific function call is from 8% to 0.32%
of total runtime.