🔥 Get rid of the webassets cache
We should rarely if ever regenerate the webassets in production and can
wait a few seconds for regeneration in development, if it means we can
get rid of a whole monkey patch and an internal use of pickle with
non-tamperproof files.

So disable the bundle component cache and switch the manifest over
to use a json file. Ensure the latter is also removed on server
start, just like the other webassets related generated files.

Better solution to what was discussed in #4339 and originally
to be solved with the now reverted 0f41a42.