✨  Add `--no-color` option to disable colored output on console (#4402)
* :sparkles: Add `--no-color` option to disable colored output on console

Closes #4396

Also removed the setting of INFO logs to white, instead it will be left as default. If there is a light-colored background (such as white) then it becomes unreadable. Can then just use whatever user has configured for their default and not worry about contrast issues.

* 🎨 Remove redundant bool wrapper

* ⬆️ Update to colorlog 6

That way we get support for NO_COLOR as well:

https://no-color.org/

* 💬 Add NO_COLOR to --no-color help

Co-authored-by: Gina Häußge <gina@octoprint.org>