Support two-digit flavors for G00 G01 G02 G03
Many CAM processors generate two-digit movement commands.  These are legal movement commands but the viewer is currently blind to them.  (Evidently slicers never generate these, or this would have already been an issue.)

This does not fully solve rendering of non-extruding toolpaths, but it enables a workaround in combination with M101 added to the preamble.