💚 Make cypress more shaky friendly
Ever since the needed upgrade, the e2e tests have
been very shaky. Let's try to up the retry count
to get this under control.

Long term I need to find out what's going on,
but currently I'm not able to even reproduce the
issue (jquery not loading). And/or maybe switch
to PlayWright if I can't get it to work reliably
anymore.