🚸 Re-enable connect dialog w/ no serial ports
Apparently disabling it if no serial ports are currently
known to the client breaks certain workflows for people.

So this is a short-term solution to fix things for them, while
technically non-sensical (why offer to connect if we don't see
anything to connect to?) So in 1.9.0 this should rather be
solved by adding some form of platform agnostic (!) auto-refresh,
either regular polling while no port is detected, or even
watchdog enabled event based refresh. See #4473.