🐛 Fix potential race condition in webassets
Could trigger when reloading page while preliminary cache
was being done, causing an internal server error
on page load, immediately fixed by a second reload.

Seems to be caused by the change in webasset handling
in 1.8.0, hence considering this a regression.

Fix consists of switching to a thread-safe in-memory
manifest storage.