⬆️ Allow ES6 code in JS files
Less than 1% of (tracked) OctoPrint client loads
happen in browsers that are still stuck on ES5.

So let's get rid of the ES5 limitation and make
life easier for anyone working on OctoPrint's JS
code.

As a test of the eslint config, login.js and
recovery.js have been switched to the use of
arrow functions.

Implements #4562