✨ Add a reverse proxy test page
Will show variables needed for URL generation and
cookie suffix generation, both as seen by client
and server, and highlight any mismatches.

Also links to the reverse proxy FAQ and includes
information on which request headers are
relevant.

Not translated, kept as simple as possible.

Not behind a login since none of this information
should be news to a client on a properly
configured proxy setup.