✨ Support writing long names to SD
If LFN_WRITE capability is reported as supported,
OctoPrint by default now will upload SD files
to the printer under their full name instead of
shortening it to DOS conventions.

The primary name as reported by the firmware
(first entry in the M20 output) is still
considered the main name and used for any file
operations like selecting or deleting. So, if
the firmware supports a DOS name here, that's
also what OctoPrint will (continue to) use for
any such operations.

Implements #4494