✨ Add new backup hooks for backup & restore
Four new hooks were added:

- octoprint.plugin.backup.before_backup
- octoprint.plugin.backup.after_backup
- octoprint.plugin.backup.before_restore
- octoprint.plugin.backup.after_restore

Those will allow plugins to perform additional
tasks right before/after backup and right before/
after restore.

This is a compromise for #4195: Instead of having
an extension point in OctoPrint to pull in
additional files, having to figure out
readability, proper allow listing to prevent
attempts to backup everything and the kitchen sink
and such, plugins CAN pull in additional stuff
they want to have in the backup into their data
folder and then on restore (or start after
restore) put it where it belongs. In an ideal
world, we would ever only want to backup stuff
that is already in the base folder anyhow, but
as #4195 demonstrates there is sometimes a need
for more.

And since these hooks are not limited to this
functionality specifically but could also be used
for other stuff (think reregistering somewhere
after restore or whatnot), I think they are a
good compromise.

Implements #4195