⏪ Bring back `octoprint_setuptools`
Apparently we have an issue when updating from an older OctoPrint install
that included that to a newer one that depends on OctoPrint-Setuptools, because
the `octoprint_setuptools` package simply won't be installed in that case.

A reinstall after updating OctoPrint makes it available, but that's not something
we can do in the field.

Long term this needs a better solution, short term we are simply going to provide
OctoPrint-Setuptools to pyproject.toml users, and a vendored copy here. It's the
only solution I see for now.

Really closes #4749