setup.py: Eliminate extra dependency variables to enable dependabot support
dependabot is currently unable to detect dependencies from `setup.py`. The reason is most likely that the effective `install_requires` variable and alike are not assigned directly, but indirectly with a set of global constants.

This commit eliminates the dedicated constants and merged their content directly into the effective `setup.py` variables, to allow dependabot detecting dependencies. Additional code comments have been added to maintain the dependency structure accomplished with the previous variables.

Signed-off-by: MichaIng <micha@dietpi.com>